Separate the definitions and implementations in header files.
