+++ CMakeLists.txt
@@ -5,6 +5,6 @@ project( OCVE )
 # add_subdirectory(egbis)
 
 find_package( OpenCV REQUIRED )
-add_executable( main main.cpp  egbis/segment-image.h egbis/segment-graph.h egbis/disjoint-set.h egbis/image.h)
+add_executable( main main.cpp egbis/segment-image.cpp egbis/segment-image.h egbis/segment-graph.cpp egbis/segment-graph.h egbis/disjoint-set.cpp egbis/disjoint-set.h egbis/image.h egbis/filter.cpp egbis/filter.h egbis/misc.cpp egbis/misc.h)
 target_link_libraries( main ${OpenCV_LIBS} )
