+++ egbis/filter.h
@@ -72,12 +72,7 @@ static image<float> *smooth(image<float> *src, float sigma) {
 }
 
 /* convolve image with gaussian filter */
-image<float> *smooth(image<uchar> *src, float sigma) {
-  image<float> *tmp = imageUCHARtoFLOAT(src);
-  image<float> *dst = smooth(tmp, sigma);
-  delete tmp;
-  return dst;
-}
+image<float> *smooth(image<uchar> *src, float sigma);
 
 /* compute laplacian */
 static image<float> *laplacian(image<float> *src) {
@@ -96,5 +91,6 @@ static image<float> *laplacian(image<float> *src) {
   }
   return dst;
 }
+
 
 #endif
