+++ egbis/segment-graph.h
@@ -31,9 +31,7 @@ typedef struct {
   int a, b;
 } edge;
 
-bool operator<(const edge &a, const edge &b) {
-  return a.w < b.w;
-}
+bool operator<(const edge &a, const edge &b);
 
 /*
  * Segment a graph
@@ -45,39 +43,6 @@ bool operator<(const edge &a, const edge &b) {
  * edges: array of edges.
  * c: constant for treshold function.
  */
-universe *segment_graph(int num_vertices, int num_edges, edge *edges, 
-			float c) { 
-  // sort edges by weight
-  std::sort(edges, edges + num_edges);
-
-  // make a disjoint-set forest
-  universe *u = new universe(num_vertices);
-
-  // init thresholds
-  float *threshold = new float[num_vertices];
-  for (int i = 0; i < num_vertices; i++)
-    threshold[i] = THRESHOLD(1,c);
-
-  // for each edge, in non-decreasing weight order...
-  for (int i = 0; i < num_edges; i++) {
-    edge *pedge = &edges[i];
-    
-    // components conected by this edge
-    int a = u->find(pedge->a);
-    int b = u->find(pedge->b);
-    if (a != b) {
-      if ((pedge->w <= threshold[a]) &&
-	  (pedge->w <= threshold[b])) {
-	u->join(a, b);
-	a = u->find(a);
-	threshold[a] = pedge->w + THRESHOLD(u->size(a), c);
-      }
-    }
-  }
-
-  // free up
-  delete threshold;
-  return u;
-}
+universe *segment_graph(int num_vertices, int num_edges, edge *edges, float c);
 
 #endif
